+++ backend/canisters/user/impl/src/lifecycle/init.rs
@@ -21,7 +21,6 @@ fn init(args: Args) {
         args.user_index_canister_id,
         args.local_user_index_canister_id,
         args.group_index_canister_id,
-        args.notifications_canister_id,
         args.escrow_canister_id,
         args.video_call_operators,
         args.username,